Split functions.js into /modules
