+++ index.js
@@ -28,11 +28,20 @@ const client = new Discord.Client({
 const config = require('./config');
 const { version } = require('./package.json');
 const emoji = require('./src/emoji');
-require('./src/functions')(client);
 
 client.config = config;
 client.version = `v${version}`;
 client.emoji = emoji;
+
+fs.readdir('./src/modules/', (err, files) => {
+  if (err) {
+    return console.error(err);
+  }
+
+  files.forEach((file) => {
+    require(`./src/modules/${file}`)(client);
+  });
+});
 
 fs.readdir('./events/', (err, files) => {
   if (err) {
