+++ src/modules/clean.js
@@ -1,0 +1,19 @@
+/* eslint-disable no-param-reassign */
+module.exports = (client) => {
+  client.clean = async (clientParam, text) => {
+    if (text && text.constructor.name === 'Promise') {
+      text = await text;
+    }
+    if (typeof evaled !== 'string') {
+      // eslint-disable-next-line global-require
+      text = require('util').inspect(text, { depth: 1 });
+    }
+
+    text = text
+      .replace(/`/g, `\`${String.fromCharCode(8203)}`)
+      .replace(/@/g, `@${String.fromCharCode(8203)}`)
+      .replace(clientParam.token, 'mfa.VkO_2G4Qv3T--NO--lWetW_tjND--TOKEN--QFTm6YGtzq9PH--4U--tG0');
+
+    return text;
+  };
+};
