+++ src/modules/permLevel.js
@@ -1,0 +1,19 @@
+module.exports = (client) => {
+  client.permLevel = (message) => {
+    let permName = 'User';
+    let permlvl = 0;
+    const permOrder = client.config.permLevels.slice(0)
+      .sort((p, c) => (p.level < c.level ? 1 : -1));
+
+    while (permOrder.length) {
+      const currentlvl = permOrder.shift();
+
+      if (currentlvl.check(client, message)) {
+        permName = currentlvl.name;
+        permlvl = currentlvl.level;
+        break;
+      }
+    }
+    return [permName, permlvl];
+  };
+};
