+++ src/modules/toProperCase.js
@@ -1,0 +1,8 @@
+module.exports = () => {
+  // eslint-disable-next-line no-extend-native
+  Object.defineProperty(String.prototype, 'toProperCase', {
+    value() {
+      return this.replace(/([^\W_]+[^\s-]*) */g, (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase());
+    },
+  });
+};
